+++ examples/vectortiles_render.groovy
@@ -13,11 +13,10 @@ import geoscript.style.Stroke
 File dir = new File("vectortiles")
 dir.mkdir()
 
-OSM osm = new OSM("Stamen Toner", [
-        "http://a.tile.stamen.com/toner",
-        "http://b.tile.stamen.com/toner",
-        "http://c.tile.stamen.com/toner",
-        "http://d.tile.stamen.com/toner"
+OSM osm = new OSM("OSM", [
+        "http://a.tile.openstreetmap.org",
+        "http://b.tile.openstreetmap.org",
+        "http://c.tile.openstreetmap.org"
 ])
 
 Pyramid pyramid = Pyramid.createGlobalMercatorPyramid()